# eskip: improve lexer performance