@szuecs 2023-11-29
> return scalars instead of token to reduce allocations

how much does this speedup?
I think the code is already ugly if you have more than 2 return values and now we have increased from 3 to 4.