@sepehrdaddev 2024-01-12
@AlexanderYastrebov seems like the fuzzer found an index out of bound in the code, pls check the workflow.
specifically the in the `ParsePredicates()` function.