@sepehrdaddev 2024-01-12
`ParsePredicates()` panics when given `///`

here is how re reproduce it:
```diff
diff --git a/eskip/eskip_test.go b/eskip/eskip_test.go
index 249dced0..f9a75270 100644
--- a/eskip/eskip_test.go
+++ b/eskip/eskip_test.go
@@ -388,6 +388,9 @@ func TestPredicateParsing(t *testing.T) {
        }, {
                title: "star notation",
                input: `*`,
+       }, {
+               title: "panics",
+               input: `///`,
        }} {
                t.Run(test.title, func(t *testing.T) {
                        p, err := ParsePredicates(test.input)
   
```