@AlexanderYastrebov 2024-01-12
@sepehrdaddev Nice one! Did you get the `///` payload from logs
```
==72==ERROR: AddressSanitizer: ABRT on unknown address 0x000000000048 (pc 0x0000005e4481 bp 0x10c000633bd0 sp 0x10c000633bb8 T0)
SCARINESS: 10 (signal)
    #0 0x5e4481 in runtime.raise.abi0 runtime/sys_linux_amd64.s:154

DEDUP_TOKEN: runtime.raise.abi0
AddressSanitizer can not provide additional info.
SUMMARY: AddressSanitizer: ABRT runtime/sys_linux_amd64.s:154 in runtime.raise.abi0
==72==ABORTING
MS: 1 CopyPart-; base unit: 994af86bef6520042698904387522e903ba5149f
0x2f,0x2f,0x2f,
///
```
?