@sepehrdaddev 2024-01-12
yes, I got it from the logs.