@AlexanderYastrebov 2024-01-12
master does not fail properly on this payload, I've created https://github.com/zalando/skipper/pull/2849