@sepehrdaddev 2024-01-12
for reference, the fuzzer found multiple payload that crashes (multiple iterations of `/`), here is another payload that causes it to crash that I found a bit interesting since it was different than others: 
` "\x2f\x2f\x00\x00\x00\xe6\xfe\x00\x00\x2f\x00\x00\x00\x00\x00\x00\x00\xe6\xfe\x00\x00\x2f\x00\x00\x00\x00"`