@sepehrdaddev 2024-01-16
I ran the fuzzer for 30 mins and it didn't find anything so I think its safe to say that its fixed.